Update the Transformers backend to use mixins (+mm pooling classes)
